+++ sixtyfps_runtime/corelib/lib.rs
@@ -34,7 +34,7 @@ pub mod component;
 pub mod items;
 pub mod model;
 pub mod properties;
-pub mod sharedarray;
+pub mod sharedvector;
 pub mod callbacks;
 pub mod string;
 
@@ -42,7 +42,7 @@ pub mod string;
 pub use string::SharedString;
 
 #[doc(inline)]
-pub use sharedarray::SharedArray;
+pub use sharedvector::SharedVector;
 
 #[doc(inline)]
 pub use graphics::Resource;
@@ -77,7 +77,7 @@ pub mod timers;
 pub fn use_modules() -> usize {
     tests::sixtyfps_mock_elapsed_time as usize
         + callbacks::ffi::sixtyfps_callback_init as usize
-        + sharedarray::ffi::sixtyfps_shared_array_empty as usize
+        + sharedvector::ffi::sixtyfps_shared_vector_empty as usize
         + layout::solve_grid_layout as usize
         + item_tree::ffi::sixtyfps_visit_item_tree as usize
         + graphics::ffi::sixtyfps_new_path_elements as usize